requirements.txt
opencv_python==4.7.0.68
pyimgur==0.6.0
pypresence==4.2.1
python-dotenv==0.21.0
